// src/generated/splice/scan-external.ts
export const spliceScanExternalMetadata = {
  "sourceId": "splice-scan-external-openapi",
  "name": "Splice Scan external API",
  "family": "scan",
  "format": "openapi",
  "stability": "stable-external",
  "artifactVersion": "0.0.1",
  "fetchedAt": "2026-04-02T20:11:16.928Z",
  "selectedSha256": "d8c78d8e2b5a69991d2e19d28f2e423dc0cc2a4e8ebd98fc707849f23e99cde3",
  "upstream": {
    "kind": "git",
    "path": "apps/scan/src/main/openapi/scan.yaml",
    "ref": "275c31b8545ba33e66576e8298521e251eb34b21",
    "repo": "https://github.com/hyperledger-labs/splice",
    "url": "https://raw.githubusercontent.com/hyperledger-labs/splice/275c31b8545ba33e66576e8298521e251eb34b21/apps/scan/src/main/openapi/scan.yaml"
  },
  "stableGenerated": true,
  "thirdPartyDir": "third_party/upstream-specs/splice-scan-external-openapi",
  "selectedSpecFile": "third_party/upstream-specs/splice-scan-external-openapi/selected.yaml"
} as const